8.0b1 Release (#2232)
* 8.0b1 release

* auto rerun flaky tests

---------

+++ coremltools/converters/mil/mil/ops/tests/iOS14/test_tensor_transformation.py
@@ -1058,7 +1058,7 @@ class TestSliceByIndex:
             )
             return x
 
-        x = np.random.rand(*INPUT_SHAPE)
+        x = np.float16(np.random.rand(*INPUT_SHAPE))
 
         # slice by index is x[begin[0]: end[0]: stride[0], begin[1]: end[1]: stride[1], ...]
         y_numpy = x[0:1:1, 0:2:1, 0:8:2, 0:12:2]
